mcp_server/requirements.txt
mcp==1.19.0
google-search-results==2.4.2
python-dotenv==1.1.1
pydantic==2.11.7
fastapi==0.116.1
fastapi-mcp==0.4.0
httpx==0.28.1 